+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+* Optional constructor arguments are lost. [#2718](https://github.com/fsprojects/fantomas/issues/2718)
+
 ## [5.2.0-alpha-012] - 2023-01-14
 
 ### Fixed